Show users in table and Show links depends of the type of user (role)
Modify the linkGroup component, to show links depends to the type of role. Add the function of show users in the table to do this, is posible with the modification of templateListUser and User to add the attribute 'role', change the name of the table to tableUser, add headers to the method getAllUsers in userService, and add the visual modification of the ListUserPage to continue the standard.