Aggregation, modification and elemination of components to list users
Elimination of tableHeader and tableRow. Aggregation of useListUsers and pagination to add logic to use pages in the table. Modify tableUser to list users in the table, modify templateListUsers to eliminate code not useful in the component and aggregation of Pagination component, and Modify the getAllUsers in the userService class to add the paggination function