Creation of disable component
Creation of a disable modal (ModalDisable) to confim the disable. Aggregation of the disable modal to disable users in the tableUser and templateListUsers component. Modification of userService to had authorization to disable in the backend. Aggregation of a disable function in useListUsers. Modify ApiResponse to add a new Promise, with a 204 response.