Creation of enable modal and modification of disable logic
Modification of component modalDisable, the modal had a Save label and is chaged to Disable, the same component is modify to had a disabled option, the admins do not have the authorization of disable an another admin. Creation of the modalEnable component, to reactivate a user, and aggregation of this enable logic to the tableUser, useListUsers and templateListUser component. And modify the enableUser in the userService to had a Promise