Aggregation of option to edit a user
Modify the userService to add a Promise in the return method (updateUser). Modify input component to delete one rounded in the visual tailwind. Creation of a buttonEdit and useEditUsers component to edit the user, with variables like user to do the future edition, the handleChange method is not finish and the handleUpdateUser needs a Notification of the status. And the templateListUsers and tableUser has modification to integrates the new edit method.