Fix views
Fix and add new implementation in the pages like: Login and List of Users. The backgroundLogo component is modify to add the enterprise logo. In the mainHeader, the icon ArrowLeftStartOnRectangleIcon is implementaded in the button of log out. The components modalDisable and modalEnable are modify to eliminated the double notification when the user is updated. The modal alerts, the timers are modify to 3 seconds. The update user section does not notify the user if this do someting wrong, the addition of a variable error is the solution, this variable is used in tableUser, templateListUsers and useEditUsers components. The useListUser component, the alerts has a error of message to show, this is changed in the disable and enable methods. And finally the authService in the method login, the method does not return a status to verify the login. In this version, the actions of Login a User, List Users, Add new Users, Edit information of Users, Disable Users and Enable Users are functional.