Implementation of List Products
Aggregation of logic to list products with pagination included. The modification into ApiResponse is to use better typescript and implements best order in the proyect, this is why the ApiPaginated was created and this api response is going to implements into userService. The Product interface was changed to delete the attribute quantity, because the diagram was changed. To list product the creation of useListProducts to manage list logic, the modification of productService, the addition of handleAxiosErrorPaginated into errorHandler, the creation of tableProduct and modification of templateListProduct to use useListProduct should have been executed. This modification. These modifications have not been tested