Modification on Form Product
The product form has changes on the logic implemented. Into the Product interface, there is a aggregation of a ErrorProductDTO the manage the errors on the products fields, the creation of validateFormProduct is a class to manage all the validation of the product and its implemented on the new useFormProduct that have all the logic of the formProduct component. Others components that has changes are the templateAddProduct and addproductPage, beacuse the page has to be the same color and fonts to all the system. And finally into the generalValidations the validateNumberLessZero and validateOnlyNumbers are created to validate when a number is less than zero and when a text is not a number.