Reactivate a product
The function of desactivate a product is necesary to have a reactivate a product, because the Cabys Code is unique, if the user wants to reintegrate a product, is not necesary to add one again. To do this the useListProduct has modifications on the header and an aggregation of a method to enable, this enable method is created in productService. The interface of the product had a aggregation of a isActive attribute. And finally the templateListProduct and tableProduct has the modification to integrate the active method explain in the useListProduct