Edit prouduct
Edit products on the table of proudcts (tableProduct), the useEditProduct is the manager to edit the product information, this useEditProduct is used in templateListProduct. The modification on productService is to type the correct type of object. And the creation of ProductUpdateDTO to manage the correct type into the backend. The method of update is not functional.