Functional disable, enable and list products
To enable and disable the products, the productServices the method of enableProduct and disableProduct has changed. The ListProduct component had a addition of a header and a template of error if the user is not logged in. And finally into the interface of a product, the codeCABYS is changed to codeCabys.